chore: rename nav2 and moveit2 folders to generic names
