+++ src/rai_bench/rai_bench/tool_calling_agent_bench/ros2_agent_tasks.py
@@ -21,7 +21,7 @@ import inflect
 from langchain_core.messages import AIMessage
 from langchain_core.messages.tool import ToolCall
 from langchain_core.tools import BaseTool
-from rai.tools.ros2.moveit2 import MoveToPointToolInput
+from rai.tools.ros2 import MoveToPointToolInput
 
 from rai_bench.tool_calling_agent_bench.agent_tasks_interfaces import (
     ROS2ToolCallingAgentTask,
