+++ src/rai_core/rai/tools/ros2/__init__.py
@@ -29,12 +29,12 @@ from .generic import (
     ROS2TopicsToolkit,
     StartROS2ActionTool,
 )
-from .moveit2.manipulation import (
+from .manipulation.custom import (
     GetObjectPositionsTool,
     MoveToPointTool,
     MoveToPointToolInput,
 )
-from .nav2.navigation import (
+from .navigation.nav2 import (
     CancelNavigateToPoseTool,
     GetNavigateToPoseFeedbackTool,
     GetNavigateToPoseResultTool,
