+++ src/rai_core/rai/tools/ros2/navigation/__init__.py
@@ -12,7 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from .navigation import (
+from .nav2 import (
     CancelNavigateToPoseTool,
     GetNavigateToPoseFeedbackTool,
     GetNavigateToPoseResultTool,
